PR comments

+++ cli/cmd/main.go
@@ -213,175 +213,181 @@ func main() {
 			},
 			{
 				Name:  "data",
-				Usage: "download data from Viam cloud",
-				UsageText: fmt.Sprintf("viam data <%s> <%s> [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s]",
-					dataFlagDestination, dataFlagDataType, dataFlagOrgIDs, dataFlagLocationIDs, dataFlagRobotID, dataFlagRobotName,
-					dataFlagPartID, dataFlagPartName, dataFlagComponentType, dataFlagComponentModel, dataFlagComponentName,
-					dataFlagStart, dataFlagEnd, dataFlagMethod, dataFlagMimeTypes, dataFlagParallelDownloads),
-				Flags: []cli.Flag{
-					&cli.StringFlag{
-						Name:     dataFlagDestination,
-						Required: true,
-						Usage:    "output directory for downloaded data",
+				Usage: "work with data",
+				Subcommands: []*cli.Command{
+					{
+						Name:  "export",
+						Usage: "download data from Viam cloud",
+						UsageText: fmt.Sprintf("viam data <%s> <%s> [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s]",
+							dataFlagDestination, dataFlagDataType, dataFlagOrgIDs, dataFlagLocationIDs, dataFlagRobotID, dataFlagRobotName,
+							dataFlagPartID, dataFlagPartName, dataFlagComponentType, dataFlagComponentModel, dataFlagComponentName,
+							dataFlagStart, dataFlagEnd, dataFlagMethod, dataFlagMimeTypes, dataFlagParallelDownloads),
+						Flags: []cli.Flag{
+							&cli.StringFlag{
+								Name:     dataFlagDestination,
+								Required: true,
+								Usage:    "output directory for downloaded data",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagDataType,
+								Required: true,
+								Usage:    "data type to be downloaded: either binary or tabular",
+							},
+							&cli.StringSliceFlag{
+								Name:     dataFlagOrgIDs,
+								Required: false,
+								Usage:    "orgs filter",
+							},
+							&cli.StringSliceFlag{
+								Name:     dataFlagLocationIDs,
+								Required: false,
+								Usage:    "locations filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagRobotID,
+								Required: false,
+								Usage:    "robot_id filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagPartID,
+								Required: false,
+								Usage:    "part_id filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagRobotName,
+								Required: false,
+								Usage:    "robot_name filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagPartName,
+								Required: false,
+								Usage:    "part_name filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagComponentType,
+								Required: false,
+								Usage:    "component_type filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagComponentModel,
+								Required: false,
+								Usage:    "component_model filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagComponentName,
+								Required: false,
+								Usage:    "component_name filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagMethod,
+								Required: false,
+								Usage:    "method filter",
+							},
+							&cli.StringSliceFlag{
+								Name:     dataFlagMimeTypes,
+								Required: false,
+								Usage:    "mime_types filter",
+							},
+							&cli.UintFlag{
+								Name:     dataFlagParallelDownloads,
+								Required: false,
+								Usage:    "number of download requests to make in parallel, with a default value of 10",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagStart,
+								Required: false,
+								Usage:    "ISO-8601 timestamp indicating the start of the interval filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagEnd,
+								Required: false,
+								Usage:    "ISO-8601 timestamp indicating the end of the interval filter",
+							},
+						},
+						Action: DataCommand,
 					},
-					&cli.StringFlag{
-						Name:     dataFlagDataType,
-						Required: true,
-						Usage:    "data type to be downloaded: either binary or tabular",
-					},
-					&cli.StringSliceFlag{
-						Name:     dataFlagOrgIDs,
-						Required: false,
-						Usage:    "orgs filter",
-					},
-					&cli.StringSliceFlag{
-						Name:     dataFlagLocationIDs,
-						Required: false,
-						Usage:    "locations filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagRobotID,
-						Required: false,
-						Usage:    "robot_id filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagPartID,
-						Required: false,
-						Usage:    "part_id filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagRobotName,
-						Required: false,
-						Usage:    "robot_name filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagPartName,
-						Required: false,
-						Usage:    "part_name filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagComponentType,
-						Required: false,
-						Usage:    "component_type filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagComponentModel,
-						Required: false,
-						Usage:    "component_model filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagComponentName,
-						Required: false,
-						Usage:    "component_name filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagMethod,
-						Required: false,
-						Usage:    "method filter",
-					},
-					&cli.StringSliceFlag{
-						Name:     dataFlagMimeTypes,
-						Required: false,
-						Usage:    "mime_types filter",
-					},
-					&cli.UintFlag{
-						Name:     dataFlagParallelDownloads,
-						Required: false,
-						Usage:    "number of download requests to make in parallel, with a default value of 10",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagStart,
-						Required: false,
-						Usage:    "ISO-8601 timestamp indicating the start of the interval filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagEnd,
-						Required: false,
-						Usage:    "ISO-8601 timestamp indicating the end of the interval filter",
+					{
+						Name:  "delete",
+						Usage: "delete data from Viam cloud",
+						UsageText: fmt.Sprintf("viam data [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s]",
+							dataFlagDataType, dataFlagOrgIDs, dataFlagLocationIDs, dataFlagRobotID, dataFlagRobotName,
+							dataFlagPartID, dataFlagPartName, dataFlagComponentType, dataFlagComponentModel, dataFlagComponentName,
+							dataFlagStart, dataFlagEnd, dataFlagMethod, dataFlagMimeTypes),
+						Flags: []cli.Flag{
+							&cli.StringFlag{
+								Name:     dataFlagDataType,
+								Required: false,
+								Usage:    "data type to be deleted: either binary or tabular",
+							},
+							&cli.StringSliceFlag{
+								Name:     dataFlagOrgIDs,
+								Required: false,
+								Usage:    "orgs filter",
+							},
+							&cli.StringSliceFlag{
+								Name:     dataFlagLocationIDs,
+								Required: false,
+								Usage:    "locations filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagRobotID,
+								Required: false,
+								Usage:    "robot_id filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagPartID,
+								Required: false,
+								Usage:    "part_id filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagRobotName,
+								Required: false,
+								Usage:    "robot_name filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagPartName,
+								Required: false,
+								Usage:    "part_name filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagComponentType,
+								Required: false,
+								Usage:    "component_type filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagComponentModel,
+								Required: false,
+								Usage:    "component_model filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagComponentName,
+								Required: false,
+								Usage:    "component_name filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagMethod,
+								Required: false,
+								Usage:    "method filter",
+							},
+							&cli.StringSliceFlag{
+								Name:     dataFlagMimeTypes,
+								Required: false,
+								Usage:    "mime_types filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagStart,
+								Required: false,
+								Usage:    "ISO-8601 timestamp indicating the start of the interval filter",
+							},
+							&cli.StringFlag{
+								Name:     dataFlagEnd,
+								Required: false,
+								Usage:    "ISO-8601 timestamp indicating the end of the interval filter",
+							},
+						},
+						Action: DeleteCommand,
 					},
 				},
-				Action: DataCommand,
-			},
-			{
-				Name:  "delete",
-				Usage: "delete data from Viam cloud",
-				UsageText: fmt.Sprintf("viam data [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s] [%s]",
-					dataFlagDataType, dataFlagOrgIDs, dataFlagLocationIDs, dataFlagRobotID, dataFlagRobotName,
-					dataFlagPartID, dataFlagPartName, dataFlagComponentType, dataFlagComponentModel, dataFlagComponentName,
-					dataFlagStart, dataFlagEnd, dataFlagMethod, dataFlagMimeTypes),
-				Flags: []cli.Flag{
-					&cli.StringFlag{
-						Name:     dataFlagDataType,
-						Required: false,
-						Usage:    "data type to be deleted: either binary or tabular",
-					},
-					&cli.StringSliceFlag{
-						Name:     dataFlagOrgIDs,
-						Required: false,
-						Usage:    "orgs filter",
-					},
-					&cli.StringSliceFlag{
-						Name:     dataFlagLocationIDs,
-						Required: false,
-						Usage:    "locations filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagRobotID,
-						Required: false,
-						Usage:    "robot_id filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagPartID,
-						Required: false,
-						Usage:    "part_id filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagRobotName,
-						Required: false,
-						Usage:    "robot_name filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagPartName,
-						Required: false,
-						Usage:    "part_name filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagComponentType,
-						Required: false,
-						Usage:    "component_type filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagComponentModel,
-						Required: false,
-						Usage:    "component_model filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagComponentName,
-						Required: false,
-						Usage:    "component_name filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagMethod,
-						Required: false,
-						Usage:    "method filter",
-					},
-					&cli.StringSliceFlag{
-						Name:     dataFlagMimeTypes,
-						Required: false,
-						Usage:    "mime_types filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagStart,
-						Required: false,
-						Usage:    "ISO-8601 timestamp indicating the start of the interval filter",
-					},
-					&cli.StringFlag{
-						Name:     dataFlagEnd,
-						Required: false,
-						Usage:    "ISO-8601 timestamp indicating the end of the interval filter",
-					},
-				},
-				Action: DeleteCommand,
 			},
 			{
 				Name:  "robots",
@@ -786,14 +792,24 @@ func main() {
 	}
 }
 
+func validateDataType(c *cli.Context) (string, error) {
+	dataType := c.String(dataFlagDataType)
+	switch dataType {
+	case dataTypeBinary, dataTypeTabular:
+		return dataType, nil
+	default:
+		return "", errors.Errorf("type must be binary or tabular, got %s", c.String("type"))
+	}
+}
+
 // DataCommand runs the data command for downloading data from the Viam cloud.
 func DataCommand(c *cli.Context) error {
-	if c.String(dataFlagDataType) != dataTypeBinary && c.String(dataFlagDataType) != dataTypeTabular {
-		return errors.Errorf("type must be binary or tabular, got %s", c.String("type"))
+	dataType, err := validateDataType(c)
+	if err != nil {
+		return err
 	}
-
-	filter := &datapb.Filter{}
-	if err := createDataFilter(c, filter); err != nil {
+	filter, err := createDataFilter(c)
+	if err != nil {
 		return err
 	}
 
@@ -802,31 +818,28 @@ func DataCommand(c *cli.Context) error {
 		return err
 	}
 
-	dataType := c.String(dataFlagDataType)
-	switch dataType {
-	case dataTypeBinary:
+	if dataType == dataTypeBinary {
 		if err := client.BinaryData(c.String(dataFlagDestination), filter, c.Uint(dataFlagParallelDownloads)); err != nil {
 			return err
 		}
-	case dataTypeTabular:
+	}
+	if dataType == dataTypeTabular {
 		if err := client.TabularData(c.String(dataFlagDestination), filter); err != nil {
 			return err
 		}
-	default:
-		return errors.Errorf("invalid data type %s", dataType)
 	}
-
 	return nil
 }
 
 // DeleteCommand runs the command for deleting data from the Viam cloud.
 func DeleteCommand(c *cli.Context) error {
-	if c.String(dataFlagDataType) != dataTypeBinary && c.String(dataFlagDataType) != dataTypeTabular {
-		return errors.Errorf("type must be binary or tabular, got %s", c.String("type"))
+	dataType, err := validateDataType(c)
+	if err != nil {
+		return err
 	}
 
-	filter := &datapb.Filter{}
-	if err := createDataFilter(c, filter); err != nil {
+	filter, err := createDataFilter(c)
+	if err != nil {
 		return err
 	}
 
@@ -835,26 +848,24 @@ func DeleteCommand(c *cli.Context) error {
 		return err
 	}
 
-	dataType := c.String(dataFlagDataType)
-	switch dataType {
-	case dataTypeBinary:
+	if dataType == dataTypeBinary {
 		if err := client.DeleteBinaryData(filter); err != nil {
 			return err
 		}
-	case dataTypeTabular:
+	}
+	if dataType == dataTypeTabular {
 		if err := client.DeleteTabularData(filter); err != nil {
 			return err
 		}
-	default:
-		return errors.Errorf("invalid data type %s", dataType)
 	}
-
 	return nil
 }
 
-func createDataFilter(c *cli.Context, filter *datapb.Filter) error {
+func createDataFilter(c *cli.Context) (*datapb.Filter, error) {
+	filter := &datapb.Filter{}
+
 	if c.String(dataFlagDataType) != dataTypeBinary && c.String(dataFlagDataType) != dataTypeTabular {
-		return errors.Errorf("type must be binary or tabular, got %s", c.String("type"))
+		return nil, errors.Errorf("type must be binary or tabular, got %s", c.String("type"))
 	}
 
 	if c.StringSlice(dataFlagOrgIDs) != nil {
@@ -897,14 +908,14 @@ func createDataFilter(c *cli.Context, filter *datapb.Filter) error {
 	if c.String(dataFlagStart) != "" {
 		t, err := time.Parse(timeLayout, c.String(dataFlagStart))
 		if err != nil {
-			return errors.Wrap(err, "error parsing start flag")
+			return nil, errors.Wrap(err, "error parsing start flag")
 		}
 		start = timestamppb.New(t)
 	}
 	if c.String(dataFlagEnd) != "" {
 		t, err := time.Parse(timeLayout, c.String(dataFlagEnd))
 		if err != nil {
-			return errors.Wrap(err, "error parsing end flag")
+			return nil, errors.Wrap(err, "error parsing end flag")
 		}
 		end = timestamppb.New(t)
 	}
@@ -914,5 +925,5 @@ func createDataFilter(c *cli.Context, filter *datapb.Filter) error {
 			End:   end,
 		}
 	}
-	return nil
+	return filter, nil
 }
